# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Control Panel API for Discord bot management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /command-logs:
    get:
      operationId: listCommandLogs
      tags: [command-logs]
      summary: List command logs with search and filter
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: userId
          in: query
          schema: { type: string }
        - name: serverId
          in: query
          schema: { type: string }
        - name: command
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [success, failure] }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Paginated command logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandLogPage"

    post:
      operationId: createCommandLog
      tags: [command-logs]
      summary: Ingest a command log from the bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandLogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandLog"
  /command-logs/export:
    get:
      operationId: exportCommandLogs
      tags: [command-logs]
      summary: Export command logs as CSV
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: query
          schema: { type: string }
        - name: serverId
          in: query
          schema: { type: string }
        - name: command
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # ─── Blacklist ──────────────────────────────────────────────────────────────
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all servers the bot is in
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: search
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema: { type: string, enum: [name, memberCount, joinedAt], default: joinedAt }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Paginated servers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerPage"
  /servers/{serverId}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a single server
      security: [{ bearerAuth: [] }]
      parameters:
        - name: serverId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Server detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"

  # ─── Users ───────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users who have interacted with the bot
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotUserPage"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile with activity history
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotUserDetail"

  # ─── Admin / Auth ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CommandLog:
      type: object
      required:
        - id
        - username
        - userId
        - serverName
        - serverId
        - command
        - status
        - timestamp
      properties:
        id: { type: integer }
        username: { type: string }
        userId: { type: string }
        serverName: { type: string }
        serverId: { type: string }
        command: { type: string }
        status: { type: string, enum: [success, failure] }
        timestamp: { type: string, format: date-time }
        errorMessage: { type: ["string", "null"] }
    CommandLogInput:
      type: object
      required: [username, userId, serverName, serverId, command, status, timestamp]
      properties:
        username: { type: string }
        userId: { type: string }
        serverName: { type: string }
        serverId: { type: string }
        command: { type: string }
        status: { type: string, enum: [success, failure] }
        timestamp: { type: string, format: date-time }
        errorMessage: { type: string }
    CommandLogPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CommandLog"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Blacklist ──────────────────────────────────────────────────────────────
...
    Server:
      type: object
      required: [id, discordId, name, memberCount, joinedAt]
      properties:
        id: { type: integer }
        discordId: { type: string }
        name: { type: string }
        memberCount: { type: integer }
        joinedAt: { type: string, format: date-time }
        iconUrl: { type: ["string", "null"] }
    ServerPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Server"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    BotUser:
      type: object
      required: [id, discordId, username, commandsUsed, lastSeen]
      properties:
        id: { type: integer }
        discordId: { type: string }
        username: { type: string }
        commandsUsed: { type: integer }
        lastSeen: { type: string, format: date-time }
        avatarUrl: { type: ["string", "null"] }
    BotUserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BotUser"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    BotUserDetail:
      type: object
      required: [user, recentActivity]
      properties:
        user:
          $ref: "#/components/schemas/BotUser"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/CommandLog"

    # ── Admin ──────────────────────────────────────────────────────────────────
...