# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Control Panel API for Discord bot management
 * OpenAPI spec version: 0.1.0
 */

export interface Server {
  id: number;
  discordId: string;
  name: string;
  memberCount: number;
  joinedAt: Date;
  /** @nullable */
  iconUrl?: string | null;
}