# lib/api-zod/src/generated/types/serverPage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Control Panel API for Discord bot management
 * OpenAPI spec version: 0.1.0
 */
import type { Server } from './server';

export interface ServerPage {
  data: Server[];
  total: number;
  page: number;
  limit: number;
}